create Button, LinkButton components

+++ src/pages/ListImagePage/ListImagePage.js
@@ -1,9 +1,12 @@
 import React, { useEffect, useCallback } from 'react';
 import { Link } from 'react-router-dom';
-import './ListImagePage.css';
+import './ListImagePage.scss';
+
 import { useSelector, useDispatch } from 'react-redux';
 import { imageListSelector} from '../../selectors'
 import { getImageList, removeImage } from '../../actions';
+import Button from '../../components/button/Button';
+import LinkButton from '../../components/link-button/LinkButton';
 
 const ListImagePage = () => {
   const imageList = useSelector(imageListSelector)
@@ -14,27 +17,27 @@ const ListImagePage = () => {
   }, [dispatch])
 
   const removeImageHandler = imageId => {
-    dispatch(removeImage(imageId))
+    if (window.confirm('Are you sure?')) {
+      dispatch(removeImage(imageId))
+    }
   }
 
   return (
     <>
       <h1>ListImagePage</h1>
-      <div className="add-image-button">
-        <Link to="/add">Add Image</Link>
-      </div>
+      <LinkButton to="/add">Add Image</LinkButton>
       <div className="gridview">
-        <div className="gridview-header">Image</div>
-        <div className="gridview-header">Actions</div>
+        <div className="cell gridview-header">Image</div>
+        <div className="cell gridview-header">Actions</div>
         {imageList.map(({ id, image }) => (
           <>
-            <div className="gridview-image-container">
+            <div className="cell gridview-image-container">
               <Link to={`/view/${id}`}>
                 <img key={id} data-id={id} src={image.contents} alt="" height="100" />
               </Link>
             </div>
-            <div>
-              <button onClick={() => removeImageHandler(id)}>Remove</button>
+            <div className="cell">
+              <Button onClick={() => removeImageHandler(id)}>Remove</Button>
             </div>
           </>
         ))}
